requirements.txt
duckdb==1.4.2
speedtest-cli==2.1.3
